STTINT-1226 read the ssl keystore from the filesystem, rather than from within Jar (#7)
We are currently reading the keystore file from the Classpath, which means that it needs to be bundled inside some jar, at build time.

The keystore file contains a private key and given that this is secret, it makes more sense for it to be read from the filesystem. This way, it can be provided at deploy time.